==> scripts/05-case.Rmd <==
---
title: "IMDB"
output: 
  flexdashboard::flex_dashboard:
    orientation: rows
    theme: journal
editor_options: 
  chunk_output_type: console
---


```{r setup, include=FALSE}
library(flexdashboard)
library(tidyverse)
imdb <- read_rds("../dados/imdb.rds")

imdb_analise <- imdb %>% 
  mutate(
    lucro = receita-orcamento,
    retorno = lucro/orcamento
  )
```

Visão geral
========================================================================

Row {data-height=110}
----------------------------------------------------------------------

### O problema

O estúdio de filmes FilmR precisa escolher um(a) diretor(a) e um(a) protagonista para o seu próximo filme. Para isso ele conduziu uma análise exploratória em uma base de dados com diversas informações de quase 4 mil filmes lançados nos Estados Unidos desde 1916. O objetivo da FilmR é identificar os diretores e atores que trariam o maior retorno financeiro para a produtora. 

Row
---

### Filmes

```{r}

imdb_analise %>%
  top_n(10, retorno) %>%
  mutate(
    #titulo = str_wrap(titulo, 15),
    titulo = fct_reorder(titulo, retorno)) %>%
  ggplot(aes(x = titulo, y = retorno)) +
  geom_col() +
  coord_flip() +
  theme_bw(20) +
  labs(x = "Título", y = "Retorno financeiro")

# library(knitr)
# 
# imdb_analise %>%
#   top_n(10, retorno) %>%
#   mutate(
#     titulo = str_wrap(titulo, 15),
#     titulo = fct_reorder(titulo, retorno)) %>%
#   select(titulo, retorno) %>% 
#   kable()

```

### Diretores

```{r}

imdb_analise %>%
  group_by(diretor) %>% 
  summarise(
    retorno_medio = mean(retorno),
    numero_de_filmes = n()
  ) %>% 
  filter(numero_de_filmes > 3) %>% 
  top_n(10, retorno_medio) %>%
  mutate(
    #titulo = str_wrap(titulo, 15),
    diretor = fct_reorder(diretor, retorno_medio)) %>%
  ggplot(aes(x = diretor, y = retorno_medio)) +
  geom_col() +
  coord_flip() +
  theme_bw(20) +
  labs(x = "Diretor", y = "Retorno financeiro")

```

### Atores

```{r}

imdb_analise %>% 
  gather(importancia, ator, ator_1, ator_2, ator_3) %>% 
  group_by(ator) %>% 
  summarise(
    retorno_medio = mean(retorno),
    numero_de_filmes = n()
  ) %>% 
  filter(numero_de_filmes > 5) %>% 
  top_n(10, retorno_medio) %>% 
  mutate(
    ator = fct_reorder(ator, retorno_medio)
  ) %>% 
  ggplot(aes(x = ator, y = retorno_medio)) +
  geom_col() +
  coord_flip() +
  theme_bw(20) +
  labs(x = "Retorno Financeiro",
       y = "Ator")

```

Row
--------------------------------------------------------------

### Relação entre nota e lucro

```{r}
imdb <- imdb %>% mutate(lucro = receita - orcamento)

imdb_analise %>% 
  filter(retorno < 100) %>% 
  ggplot(aes(x = nota_imdb, y = retorno)) +
  geom_point() +
  theme_bw(20) +
  labs(x = "Nota", y = "Retorno Financeiro")


```

> Parece que se a nota é inferiror a uns 7, ela não é muito relacionada ao lucro do filme.
Mas em geral filmes com notas bem altas lucram mais do que os outros.

### Relação entre o orçamento e lucro

```{r}

imdb_analise %>% 
  filter(retorno < 100) %>% 
  ggplot(aes(x = orcamento, y = lucro)) +
  geom_point() +
  theme_bw(20) +
  labs(x = "Orçamento", y = "Lucro")

```

---

> Filmes com maior orçamento têm maior possibilidade de prejuizo. Essas duas variáveis não são fortemente correlacionadas.

Gêneros
==============================================================


Row
---------------------------------------------------------------

### Lucro por gênero

```{r}

imdb_analise %>% 
  group_by(generos) %>% 
  # summarise(
  #   lucro_medio = mean(lucro, na.rm = TRUE),
  #   total_de_filmes = n()
  # ) %>% 
  filter(n() > 30) %>% 
  ungroup() %>% 
  mutate(generos = fct_reorder(generos, lucro, .fun = median)) %>% 
  ggplot(aes(x = generos, y = lucro)) +
  geom_boxplot() +
  coord_flip() +
  theme_bw(20)

```

> O lucro mediano dos filmes parece ser bem parecido entre os principais gêneros. No entanto, parece que alguns gêneros tem bem mais filmes que estouraram.

### Filmes que estouraram

```{r}
calcula_valor_corte <- function(x) {
 quartis <- quantile(x, c(0.25,0.5,0.75), na.rm = TRUE)
 quartis[2] + 2*(quartis[3] - quartis[1])
}

```

> Vemos que os filmes de animação tem a maior chance de "estourar" seguidos por aventura e ação.

Diretores {data-orientation=columns}
============================================

Column
-------------------------------------------

### Diretores com mais filmes

Sabemos que o estúdio quer contratar um(a) diretor(a) experiente, para ter menos riscos. Por isso vamos listar os diretores com maior número de filmes realizados.

```{r}

# graficos de barras
# grafico de diretor com mais filmes

imdb_analise %>% 
  group_by(diretor) %>% 
  summarise(
    nota_media = mean(nota_imdb),
    numero_de_filmes = n()
  ) %>% 
  top_n(10, numero_de_filmes) %>% 
  ungroup() %>% 
  mutate(
    diretor = fct_reorder(diretor, numero_de_filmes)
  ) %>% 
  ggplot(aes(x = diretor, y = numero_de_filmes)) + 
  geom_col() +
  theme_bw(20) +
  coord_flip()

```

Column {.tabset}
--------------------------------

### Notas no IMDB

```{r fig.height=7}

# geom_point
# diretores com maiores médias de nota do imdb

imdb_analise %>% 
  group_by(diretor) %>% 
  summarise(
    nota_media = mean(nota_imdb),
    numero_de_filmes = n()
  ) %>% 
  filter(numero_de_filmes > 3) %>% 
  top_n(10, nota_media) %>% 
  mutate(diretor = fct_reorder(diretor, nota_media)) %>% 
  ggplot(aes(x = diretor, y = nota_media)) +
  geom_point() +
  theme_bw(15) +
  coord_flip() +
  labs(x = "Diretor", y = "Nota média no IMDB")
``` 

### Lucro

```{r fig.height=7}

# boxplot
# diretores com maiores lucros


imdb_analise %>% 
  filter(!is.na(diretor)) %>% 
  group_by(diretor) %>% 
  # summarise(
  #   nota_media = mean(nota_imdb),
  #   numero_de_filmes = n()
  # ) %>% 
  filter(n() > 10, retorno < 30) %>% 
  ungroup() %>% 
  #top_n(10, nota_media) %>% 
  mutate(diretor = fct_reorder(diretor, retorno, .fun = median)) %>% 
  ggplot(aes(x = diretor, y = retorno)) +
  geom_boxplot() +
  theme_bw(15) +
  coord_flip() +
  labs(x = "Diretor", y = "Retorno")
```

Atores {data-orientation=columns}
============================================

Column
--------------------------------------------

### Principais atores

O estúdio prefere um ator com certa experiência, que pode ser medida pelo número de filmes. No entanto, não quer deixar a qualidade de lado - quer os atores de participaram com filmes que tiveram boas avaliações.

```{r}

# vizualizacao dos atores com boas avaliações

# imdb_analise %>% 
#   gather(importancia, ator, ator_1, ator_2, ator_3) %>% 
#   filter(nota_imdb > 7) %>% 
#   count(ator) %>% 
#   top_n(10, n) %>% 
#   mutate(ator = fct_reorder(ator, n)) %>% 
#   ggplot(aes(x = ator, y = n)) + 
#   geom_col() +
#   coord_flip() +
#   theme_bw(20)

imdb_analise %>% 
  gather(importancia, ator, ator_1, ator_2, ator_3) %>% 
  #filter(nota_imdb > 7) %>% 
  group_by(ator) %>% 
  summarise(
    filmes_com_nota_alta = sum(nota_imdb > 7, na.rm = TRUE)
  ) %>% 
  top_n(10, filmes_com_nota_alta) %>% 
  mutate(ator = fct_reorder(ator, filmes_com_nota_alta)) %>% 
  ggplot(aes(x = ator, y = filmes_com_nota_alta)) + 
  geom_col() +
  coord_flip() +
  theme_bw(20)
  

```

# Column
--------------------------------------

### Relação Atores e Diretores

Sabemos que em geral os diretores possuem preferência por alguns atores. Por isso,
listamos para cada diretor que consideramos anteriormente, o ator que mais trabalhou em conjunto.

```{r}

# vizualizacao do ator que trabalhou mais vezes com cada diretor
# pode ter só os 10 diretores com mais filmes

imdb_analise %>% 
  group_by(diretor) %>% 
  filter(n() > 12) %>% 
  gather(importancia, ator, ator_1, ator_2, ator_3) %>% 
  count(diretor, ator) %>% 
  group_by(diretor) %>% 
  filter(n == max(n)) %>% 
  knitr::kable()
```
